requirements.txt
# Core
pandas==2.2.0
numpy==1.26.2
python-dotenv==1.0.0

# Azure
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0

# Spotify
spotipy==2.23.0

# Data formats
pyarrow==15.0.0

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0